+++ projects/ng-virtual-list/src/lib/ng-virtual-list.component.ts
@@ -21,8 +21,7 @@ import {
   MAX_VELOCITY_FOR_SCROLL_QUALITY_OPTIMIZATION_LVL2, PREPARE_ITERATIONS_FOR_UPDATE_ITEMS, PREPARATION_REUPDATE_LENGTH_FOR_UPDATE_ITEMS,
   PREPARE_ITERATIONS_FOR_COLLAPSE_ITEMS, PREPARATION_REUPDATE_LENGTH_FOR_COLLAPSE_ITEMS, MAX_NUMBERS_OF_SKIPS_FOR_QUALITY_OPTIMIZATION_LVL1,
   DEFAULT_SCROLLING_SETTINGS, DEFAULT_SNAP_TO_ITEM, DEFAULT_SNAP_TO_ITEM_ALIGN, VIEWPORT, DEFAULT_MOTION_BLUR, DISABLED,
-  DEFAULT_MAX_MOTION_BLUR,
-  DEFAULT_SCROLLING_ONE_BY_ONE,
+  DEFAULT_MAX_MOTION_BLUR, DEFAULT_SCROLLING_ONE_BY_ONE, DEFAULT_MOTION_BLUR_ENABLED,
 } from './const';
 import {
   IRenderVirtualListItem, IVirtualListCollection, IVirtualListItem, IVirtualListItemConfigMap,
@@ -684,11 +683,11 @@ export class NgVirtualListComponent implements OnDestroy {
   scrollingOneByOne = input<boolean>(DEFAULT_SCROLLING_ONE_BY_ONE, { ...this._scrollingOneByOneOptions });
 
   private _motionBlurOptions = {
-    transform: (v: any) => {
-      const valid = validateFloat(v) || (v === DISABLED);
+    transform: (v: number) => {
+      const valid = validateFloat(v);
 
       if (!valid) {
-        console.error('The "motionBlur" parameter must be of type `number` or `disabled`.');
+        console.error('The "motionBlur" parameter must be of type `number`.');
         return DEFAULT_MOTION_BLUR;
       }
       return v;
@@ -698,7 +697,7 @@ export class NgVirtualListComponent implements OnDestroy {
   /**
    * Motion blur effect. The default value is `0.25`.
    */
-  motionBlur = input<number | 'disabled'>(DEFAULT_MOTION_BLUR, { ...this._motionBlurOptions });
+  motionBlur = input<number>(DEFAULT_MOTION_BLUR, { ...this._motionBlurOptions });
 
   private _maxMotionBlurOptions = {
     transform: (v: number) => {
@@ -716,6 +715,23 @@ export class NgVirtualListComponent implements OnDestroy {
    * Maximum motion blur effect. The default value is `20`.
    */
   maxMotionBlur = input<number>(DEFAULT_MAX_MOTION_BLUR, { ...this._maxMotionBlurOptions });
+
+  private _motionBlurEnabledOptions = {
+    transform: (v: boolean) => {
+      const valid = validateBoolean(v);
+
+      if (!valid) {
+        console.error('The "motionBlurEnabled" parameter must be of type `boolean`.');
+        return DEFAULT_MOTION_BLUR_ENABLED;
+      }
+      return v;
+    },
+  } as any;
+
+  /**
+   * Determines whether to apply motion blur or not. The default value is `false`.
+   */
+  motionBlurEnabled = input<boolean>(DEFAULT_MOTION_BLUR_ENABLED, { ...this._motionBlurEnabledOptions });
 
   private _animationParamsOptions = {
     transform: (v: IAnimationParams) => {